+++ BBMOD_CLI/include/BBMOD/Model.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <BBMOD/common.hpp>
 #include <BBMOD/Config.hpp>
 #include <BBMOD/VertexFormat.hpp>
 #include <BBMOD/Node.hpp>
@@ -30,11 +31,11 @@ struct SModel
 	
 	std::vector<SMesh*> Meshes;
 
-	size_t NodeCount = 0;
+	uint32_t NodeCount = 0;
 
 	SNode* RootNode = nullptr;
 
-	size_t BoneCount = 0;
+	uint32_t BoneCount = 0;
 
 	std::vector<SBone*> Skeleton;
