+++ BBMOD_CLI/src/main.cpp
@@ -98,7 +98,7 @@ int main(int argc, const char* argv[])
 			{
 				auto o = match[1];
 				bool bValue = (match[2] == "true");
-				size_t iValue = (size_t)strtol(match[2].str().c_str(), (char**)NULL, 10);
+				uint32_t iValue = (uint32_t)strtol(match[2].str().c_str(), (char**)NULL, 10);
 
 				if (o == "-lh" || o == "--left-handed")
 				{
